# application/src/main/java/com/wgblackmon/aihealthcare/web/dto/VariantRequest.java
package com.wgblackmon.aihealthcare.web.dto;

/**
 * Request DTO for creating a new prompt variant.
 *
 * @param variantId    unique identifier for the variant
 * @param name         human-readable label
 * @param templateText the full prompt template with placeholders
 * @param description  free-text description of the experiment (optional)
 *
 * @author  Bill Blackmon
 * @version 1.0
 * @since   2026-04-18
 * @updated 2026-04-18
 */
public record VariantRequest(
        String variantId,
        String name,
        String templateText,
        String description
) {}